viktor-app/requirements.txt
viktor
openai-agents
python-dotenv
pydantic
requests
